requirements.txt
django == 5.2.3
djangorestframework == 3.16.0
django-extensions
requests == 2.32.4
django-cors-headers == 4.7.0
mysqlclient == 2.2.7
drf-spectacular == 0.29.0

#===MACHINE LEARNING===
flask==2.3.3
flask-cors==4.0.0
setuptools==58.1.0
sqlparse==0.5.5
threadpoolctl==3.6.0
typing_extensions==4.15.0
numpy==2.0.2
scipy==1.13.1
joblib==1.5.3
cython==0.29.33
scikit-learn==1.6.1
asgiref == 3.11.1
httpx==0.24.1
supabase==2.1.1
gotrue==2.8.0
python-dotenv==1.2.1

